pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mkreports"
authors = [{name = "Holger Hoefling", email = "hhoeflin@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.8"

dependencies = [
    "mkdocs >=1.2.3",
    "mkdocs-material >=7.3.6",
    "mdutils >=1.3.1",
    "PyYAML >=6.0",
    "immutabledict >=2.2.1",
    "tabulate >=0.8.9",
    "deepmerge >=0.3.0",
    "python-frontmatter >=1.0.0",
    "anytree >=2.8.0",
    "intervaltree >=3.1.0",
    "more-itertools >=8.12.0",
    "typer >=0.4.0",
    "mkdocstrings[python] >=0.17.0",
    "GitPython >=3.1.26",
    "parse >=1.19.0",
    "importlib-resources >=5.4.0; python_version <= 3.8",
    "sh",
    "ipython >=7.30.1",
    "attrs",
    "overrides",
]

[project.optional-dependencies]
dev = [
    "pudb >=2021.2.2",
    "bpython >=0.22.1",
    "pytest >=6.2.5",
    "pytest-pudb >=0.7.0",
    "black",
    "flakeheaven",
    "types-PyYAML",
    "pandas-stubs",
    "types-Pillow",
    "pydocstyle",
]
content = [
    "pandas",
    "plotnine >=0.8.0",
    "black",
    "matplotlib >=3.4.3",
    "plotly >=5.5.0",
    "seaborn >=0.11.2",
    "altair >=4.1.0",
]

[project.urls]
Home = "https://hhoeflin.github.io/mkreports/"
Repository = "https://github.com/hhoeflin/mkreports"

[project.scripts]
mkreports = "mkreports.cli:app"

[tool.pyright]
include=['mkreports']
typeCheckingMode='off'

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.flakeheaven]
max_line_length = 88

[tool.pydocstyle]
add_ignore = ["D105"]
ignore_decorators = ["overrides"]
